Heartbeat: align canonical encoding with Weft Protocol v0.1 §1
The spec permits the prototype's JSON+BLAKE2b profile but requires it not
create incompatible meanings. Implement the determinism rules stdlib can
honor; document the rest for the Rust port.

- Domain-separated hashing: digest = BLAKE2b("decima:v0.1:" || kind || 0x00
  || bytes), with disjoint event-id and cell-id spaces. weft.events()
  verifies under the same kind.
- No floats in signed content: attenuated budgets stay int.
- NFC-normalize human text at the say() boundary; canonical bytes are UTF-8.
- Add heartbeat/PROFILE.md pinning the exact delta vs v0.1 (BLAKE3-256,
  deterministic CBOR, kind-prefixed base32 ids, full AuthorizationProof,
  effect receipts, ed25519) and why the big items defer (no BLAKE3/CBOR in
  stdlib; zero-dependency constraint).
- CONTRIBUTING: commit before pull --rebase (rebase refuses a dirty tree).

smoke.py still green; NFC convergence verified (decomposed == precomposed).

+++ heartbeat/decima/capability.py
@@ -127,7 +127,8 @@ def attenuate(parent_content: dict, stricter: dict, parent_id: str,
     caveats = dict(parent_content.get("caveats", {}))
     for k, v in stricter.items():
         if k == "budget":
-            caveats["budget"] = min(float(v), float(caveats.get("budget", v)))
+            # ints only — floats are forbidden in canonical/hashed content (§1)
+            caveats["budget"] = min(int(v), int(caveats.get("budget", v)))
         else:
             caveats[k] = v  # adding a constraint (e.g. requires_approval) only narrows
     return capability_content(
